[refactor] 영수증 스타일 분리 및 회비 메모 ui 수정
