+++ src/components/Receipt/ReceiptMain.tsx
@@ -1,10 +1,9 @@
-import styled from 'styled-components';
 import { useContext, useState } from 'react';
 import ReactModal from 'react-modal';
-import theme from '@/styles/theme';
 import ReceiptInfo from '@/components/Receipt/ReceiptInfo';
 import { DuesContext } from '@/service/DuesContext';
 import DuesAPI from '@/service/DuesAPI';
+import * as S from '@/styles/receipt/ReceiptMain.styled';
 
 interface ReceiptProps {
   id: number;
@@ -17,91 +16,6 @@ interface ReceiptProps {
 interface GroupedByMonth {
   [key: number]: ReceiptProps[];
 }
-
-const StyledReceipt = styled.div`
-  width: 370px;
-  height: calc(var(--vh, 1vh) * 100);
-  font-family: ${theme.font.family.pretendard_regular};
-`;
-
-const Line = styled.div`
-  border: 1px solid;
-  color: #4d4d4d;
-  width: 88%;
-  margin: 15px 6% 0 6%;
-  transform: scaleY(0.2);
-`;
-
-const StyledMonth = styled.div`
-  font-family: ${theme.font.family.pretendard_semiBold};
-  font-size: 18px;
-  margin: 15px 0 0 6%;
-`;
-
-const ScrollContainer = styled.div`
-  display: flex;
-  width: 88%;
-  margin: 15px 6% 0 6%;
-  overflow-x: auto;
-  &::-webkit-scrollbar {
-    height: 8px;
-  }
-
-  &::-webkit-scrollbar-thumb {
-    border-radius: 4px;
-  }
-`;
-
-const GridItem = styled.div`
-  flex: 0 0 auto;
-  margin-right: 10px;
-  padding: 0;
-  background-color: ${theme.color.grayScale.gray18};
-  width: 56%;
-  height: 124px;
-  color: #fff;
-  border-radius: 10px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 14px;
-  white-space: nowrap;
-  cursor: pointer;
-  position: relative;
-  &:last-child {
-    margin-right: 0;
-  }
-`;
-
-const OpenModalButton = styled.button`
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: transparent;
-  border: none;
-`;
-
-const GridItemImage = styled.embed`
-  width: 100%;
-  height: 100%;
-  border-radius: 10px;
-  object-fit: cover;
-  scroll: no;
-  overflow: hidden;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-`;
-
-const ModalImage = styled.embed`
-  width: 100%;
-  height: 100%;
-  object-fit: contain;
-`;
 
 const ReceiptMain: React.FC = () => {
   const { duesData } = useContext(DuesContext);
@@ -141,11 +55,11 @@ const ReceiptMain: React.FC = () => {
   }
 
   return (
-    <StyledReceipt>
+    <S.StyledReceipt>
       <DuesAPI />
       {months.map((month) => (
         <div key={month}>
-          <StyledMonth>{month}월</StyledMonth>
+          <S.StyledMonth>{month}월</S.StyledMonth>
           {groupedByMonth[month] ? (
             groupedByMonth[month].map((receipt) => (
               <div key={receipt.id}>
@@ -154,27 +68,27 @@ const ReceiptMain: React.FC = () => {
                   date={new Date(receipt.date).toLocaleDateString('ko-KR')}
                   memo={receipt.description}
                 />
-                <ScrollContainer>
+                <S.ScrollContainer>
                   {receipt.images.length > 0 ? (
                     receipt.images.map((image, index) => (
-                      <GridItem onClick={() => openModal(image)}>
-                        <GridItemImage
+                      <S.GridItem onClick={() => openModal(image)}>
+                        <S.GridItemImage
                           src={image}
                           title={`영수증 사진 ${index + 1}`}
                         />
-                        <OpenModalButton onClick={() => openModal(image)} />
-                      </GridItem>
+                        <S.OpenModalButton onClick={() => openModal(image)} />
+                      </S.GridItem>
                     ))
                   ) : (
                     <div> </div>
                   )}
-                </ScrollContainer>
+                </S.ScrollContainer>
               </div>
             ))
           ) : (
             <div> </div>
           )}
-          <Line />
+          <S.Line />
         </div>
       ))}
       <ReactModal
@@ -200,9 +114,9 @@ const ReceiptMain: React.FC = () => {
           },
         }}
       >
-        <ModalImage src={selectedImage} title="영수증 큰 이미지" />
+        <S.ModalImage src={selectedImage} title="영수증 큰 이미지" />
       </ReactModal>
-    </StyledReceipt>
+    </S.StyledReceipt>
   );
 };
