+++ vite.config.js
@@ -1,4 +1,4 @@
-import path from 'path';  // path 모듈 추가
+import path from 'path'; // path 모듈 추가
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import svgr from 'vite-plugin-svgr';
@@ -6,8 +6,8 @@ import svgr from 'vite-plugin-svgr';
 export default defineConfig({
   plugins: [react(), svgr()],
   server: {
-		port: 3000,
-	},
+    port: 3000,
+  },
   build: {
     outDir: 'build',
   },
